test: extract shared plan/session streaming helpers
Move duplicated SSE assertion helpers into tests/helpers/streaming.ts
and consolidate create/retry boundary fixtures and request builders in
stream-session-test-helpers.ts so API and session stream specs stay
focused on behavior without changing production code.
