+++ tests/integration/features/plans/session/respond-retry-stream.spec.ts
@@ -1,121 +1,44 @@
-import type {
-  GenerationAttemptResult,
-  ProcessGenerationInput,
-} from '@/features/plans/lifecycle/types';
+import type { ProcessGenerationInput } from '@/features/plans/lifecycle/types';
 import {
   createPlanGenerationSessionBoundary,
   PLAN_RETRY_RESERVATION_ALLOWED_STATUSES,
-  type RespondRetryStreamArgs,
-  type RetryPlanGenerationPlanSnapshot,
 } from '@/features/plans/session/plan-generation-session';
 import * as streamCleanup from '@/features/plans/session/stream-cleanup';
-import { ensureUser } from '@tests/helpers/db/users';
-import {
-  buildMockProcessLifecycle,
-  type MockProcessLifecycleHandle,
-} from './stream-session-test-helpers';
 import {
   findStreamingEvent,
   readStreamingResponse,
 } from '@tests/helpers/streaming';
-import { buildTestAuthUserId, buildTestEmail } from '@tests/helpers/testIds';
 import { describe, expect, it, vi } from 'vitest';
-import { db } from '@supabase/service-role';
-
-const SUCCESS_ATTEMPT_RESULT: GenerationAttemptResult = {
-  status: 'generation_success',
-  data: {
-    modules: [
-      {
-        title: 'Retry Module',
-        estimatedMinutes: 90,
-        tasks: [
-          { title: 'Retry Task A', estimatedMinutes: 30 },
-          { title: 'Retry Task B', estimatedMinutes: 60 },
-        ],
-      },
-    ],
-    metadata: {
-      provider: 'mock',
-      model: 'mock-model',
-      usage: { promptTokens: 1, completionTokens: 1, totalTokens: 2 },
-    },
-    durationMs: 5,
-  },
-};
-
-const BASE_PLAN_SNAPSHOT: RetryPlanGenerationPlanSnapshot = {
-  topic: 'Retry Topic',
-  skillLevel: 'intermediate',
-  weeklyHours: 6,
-  learningStyle: 'mixed',
-  startDate: '2030-01-01',
-  deadlineDate: '2030-06-01',
-  origin: 'ai',
-};
-
-function buildRetryRequest(planId: string, signal?: AbortSignal): Request {
-  return new Request(`http://localhost/api/v1/plans/${planId}/retry`, {
-    method: 'POST',
-    ...(signal ? { signal } : {}),
-  });
-}
-
-interface BuildArgsInput {
-  req: Request;
-  authUserId: string;
-  internalUserId: string;
-  planId?: string;
-  plan?: RetryPlanGenerationPlanSnapshot;
-  responseHeaders?: HeadersInit;
-  requestId?: string;
-}
-
-function buildArgs(input: BuildArgsInput): RespondRetryStreamArgs {
-  return {
-    req: input.req,
-    authUserId: input.authUserId,
-    internalUserId: input.internalUserId,
-    planId: input.planId ?? 'plan_boundary_retry',
-    plan: input.plan ?? { ...BASE_PLAN_SNAPSHOT },
-    tierDb: db,
-    ...(input.responseHeaders
-      ? { responseHeaders: input.responseHeaders }
-      : {}),
-    ...(input.requestId !== undefined ? { requestId: input.requestId } : {}),
-  };
-}
-
-async function setupUser(scenario: string): Promise<{
-  authUserId: string;
-  internalUserId: string;
-}> {
-  const authUserId = buildTestAuthUserId(scenario);
-  const internalUserId = await ensureUser({
-    authUserId,
-    email: buildTestEmail(authUserId),
-    subscriptionTier: 'pro',
-  });
-  return { authUserId, internalUserId };
-}
+import {
+  BASE_RETRY_PLAN_SNAPSHOT,
+  buildMockProcessLifecycle,
+  buildRetryStreamArgs,
+  buildRetryStreamRequest,
+  setupPlanSessionUser,
+  SUCCESS_RETRY_ATTEMPT_RESULT,
+  type MockProcessLifecycleHandle,
+} from './stream-session-test-helpers';
 
 describe('PlanGenerationSessionBoundary.respondRetryStream', () => {
   it('emits plan_start with retry attempt number then complete on success', async () => {
-    const fake = buildMockProcessLifecycle(async () => SUCCESS_ATTEMPT_RESULT, {
-      topic: BASE_PLAN_SNAPSHOT.topic,
-    });
+    const fake = buildMockProcessLifecycle(
+      async () => SUCCESS_RETRY_ATTEMPT_RESULT,
+      {
+        topic: BASE_RETRY_PLAN_SNAPSHOT.topic,
+      },
+    );
     const createLifecycleService = vi.fn(() => fake.service);
     const boundary = createPlanGenerationSessionBoundary({
       createLifecycleService,
     });
 
-    const { authUserId, internalUserId } = await setupUser(
+    const { authUserId, internalUserId } = await setupPlanSessionUser(
       'boundary-retry-success',
     );
-    const req = buildRetryRequest('plan_retry_success');
+    const req = buildRetryStreamRequest('plan_retry_success');
 
     const response = await boundary.respondRetryStream(
-      buildArgs({
+      buildRetryStreamArgs({
         req,
         authUserId,
         internalUserId,
@@ -134,7 +57,7 @@ describe('PlanGenerationSessionBoundary.respondRetryStream', () => {
     expect(planStart?.data).toMatchObject({
       planId: 'plan_retry_success',
       attemptNumber: 2,
-      topic: BASE_PLAN_SNAPSHOT.topic,
+      topic: BASE_RETRY_PLAN_SNAPSHOT.topic,
     });
     expect(complete?.data).toMatchObject({
       planId: 'plan_retry_success',
@@ -156,13 +79,13 @@ describe('PlanGenerationSessionBoundary.respondRetryStream', () => {
       createLifecycleService: () => fake.service,
     });
 
-    const { authUserId, internalUserId } = await setupUser(
+    const { authUserId, internalUserId } = await setupPlanSessionUser(
       'boundary-retry-retryable',
     );
 
     const response = await boundary.respondRetryStream(
-      buildArgs({
-        req: buildRetryRequest('plan_retry_retryable'),
+      buildRetryStreamArgs({
+        req: buildRetryStreamRequest('plan_retry_retryable'),
         authUserId,
         internalUserId,
         planId: 'plan_retry_retryable',
@@ -192,13 +115,13 @@ describe('PlanGenerationSessionBoundary.respondRetryStream', () => {
       createLifecycleService: () => fake.service,
     });
 
-    const { authUserId, internalUserId } = await setupUser(
+    const { authUserId, internalUserId } = await setupPlanSessionUser(
       'boundary-retry-reqid',
     );
 
     const response = await boundary.respondRetryStream(
-      buildArgs({
-        req: buildRetryRequest('plan_retry_reqid'),
+      buildRetryStreamArgs({
+        req: buildRetryStreamRequest('plan_retry_reqid'),
         authUserId,
         internalUserId,
         planId: 'plan_retry_reqid',
@@ -224,13 +147,13 @@ describe('PlanGenerationSessionBoundary.respondRetryStream', () => {
       createLifecycleService: () => fake.service,
     });
 
-    const { authUserId, internalUserId } = await setupUser(
+    const { authUserId, internalUserId } = await setupPlanSessionUser(
       'boundary-retry-permanent',
     );
 
     const response = await boundary.respondRetryStream(
-      buildArgs({
-        req: buildRetryRequest('plan_retry_permanent'),
+      buildRetryStreamArgs({
+        req: buildRetryStreamRequest('plan_retry_permanent'),
         authUserId,
         internalUserId,
         planId: 'plan_retry_permanent',
@@ -255,19 +178,19 @@ describe('PlanGenerationSessionBoundary.respondRetryStream', () => {
       async () => {
         throw new Error('retry boom');
       },
-      { topic: BASE_PLAN_SNAPSHOT.topic },
+      { topic: BASE_RETRY_PLAN_SNAPSHOT.topic },
     );
     const boundary = createPlanGenerationSessionBoundary({
       createLifecycleService: () => fake.service,
     });
 
-    const { authUserId, internalUserId } = await setupUser(
+    const { authUserId, internalUserId } = await setupPlanSessionUser(
       'boundary-retry-unhandled',
     );
 
     const response = await boundary.respondRetryStream(
-      buildArgs({
-        req: buildRetryRequest('plan_retry_unhandled'),
+      buildRetryStreamArgs({
+        req: buildRetryStreamRequest('plan_retry_unhandled'),
         authUserId,
         internalUserId,
         planId: 'plan_retry_unhandled',
@@ -298,19 +221,22 @@ describe('PlanGenerationSessionBoundary.respondRetryStream', () => {
         controller.abort();
         throw new DOMException('Client disconnected', 'AbortError');
       },
-      { topic: BASE_PLAN_SNAPSHOT.topic },
+      { topic: BASE_RETRY_PLAN_SNAPSHOT.topic },
     );
     const boundary = createPlanGenerationSessionBoundary({
       createLifecycleService: () => fake.service,
     });
 
-    const { authUserId, internalUserId } = await setupUser(
+    const { authUserId, internalUserId } = await setupPlanSessionUser(
       'boundary-retry-disconnect',
     );
 
     const response = await boundary.respondRetryStream(
-      buildArgs({
-        req: buildRetryRequest('plan_retry_disconnect', controller.signal),
+      buildRetryStreamArgs({
+        req: buildRetryStreamRequest(
+          'plan_retry_disconnect',
+          controller.signal,
+        ),
         authUserId,
         internalUserId,
         planId: 'plan_retry_disconnect',
@@ -325,20 +251,23 @@ describe('PlanGenerationSessionBoundary.respondRetryStream', () => {
   });
 
   it('passes responseHeaders through to the streaming Response', async () => {
-    const fake = buildMockProcessLifecycle(async () => SUCCESS_ATTEMPT_RESULT, {
-      topic: BASE_PLAN_SNAPSHOT.topic,
-    });
+    const fake = buildMockProcessLifecycle(
+      async () => SUCCESS_RETRY_ATTEMPT_RESULT,
+      {
+        topic: BASE_RETRY_PLAN_SNAPSHOT.topic,
+      },
+    );
     const boundary = createPlanGenerationSessionBoundary({
       createLifecycleService: () => fake.service,
     });
 
-    const { authUserId, internalUserId } = await setupUser(
+    const { authUserId, internalUserId } = await setupPlanSessionUser(
       'boundary-retry-headers',
     );
 
     const response = await boundary.respondRetryStream(
-      buildArgs({
-        req: buildRetryRequest('plan_retry_headers'),
+      buildRetryStreamArgs({
+        req: buildRetryStreamRequest('plan_retry_headers'),
         authUserId,
         internalUserId,
         planId: 'plan_retry_headers',
@@ -361,21 +290,21 @@ describe('PlanGenerationSessionBoundary.respondRetryStream', () => {
     const fake = buildMockProcessLifecycle(
       async (input) => {
         captured.push(input);
-        return SUCCESS_ATTEMPT_RESULT;
+        return SUCCESS_RETRY_ATTEMPT_RESULT;
       },
-      { topic: BASE_PLAN_SNAPSHOT.topic },
+      { topic: BASE_RETRY_PLAN_SNAPSHOT.topic },
     );
     const boundary = createPlanGenerationSessionBoundary({
       createLifecycleService: () => fake.service,
     });
 
-    const { authUserId, internalUserId } = await setupUser(
+    const { authUserId, internalUserId } = await setupPlanSessionUser(
       'boundary-retry-allowed-statuses',
     );
 
     const response = await boundary.respondRetryStream(
-      buildArgs({
-        req: buildRetryRequest('plan_retry_allowed'),
+      buildRetryStreamArgs({
+        req: buildRetryStreamRequest('plan_retry_allowed'),
         authUserId,
         internalUserId,
         planId: 'plan_retry_allowed',
@@ -394,9 +323,9 @@ describe('PlanGenerationSessionBoundary.respondRetryStream', () => {
     const builtFakes: MockProcessLifecycleHandle[] = [];
     const createLifecycleService = vi.fn(() => {
       const next = buildMockProcessLifecycle(
-        async () => SUCCESS_ATTEMPT_RESULT,
+        async () => SUCCESS_RETRY_ATTEMPT_RESULT,
         {
-          topic: BASE_PLAN_SNAPSHOT.topic,
+          topic: BASE_RETRY_PLAN_SNAPSHOT.topic,
         },
       );
       builtFakes.push(next);
@@ -406,22 +335,22 @@ describe('PlanGenerationSessionBoundary.respondRetryStream', () => {
       createLifecycleService,
     });
 
-    const { authUserId, internalUserId } = await setupUser(
+    const { authUserId, internalUserId } = await setupPlanSessionUser(
       'boundary-retry-factory',
     );
 
     const responses = await Promise.all([
       boundary.respondRetryStream(
-        buildArgs({
-          req: buildRetryRequest('plan_retry_factory_a'),
+        buildRetryStreamArgs({
+          req: buildRetryStreamRequest('plan_retry_factory_a'),
           authUserId,
           internalUserId,
           planId: 'plan_retry_factory_a',
         }),
       ),
       boundary.respondRetryStream(
-        buildArgs({
-          req: buildRetryRequest('plan_retry_factory_b'),
+        buildRetryStreamArgs({
+          req: buildRetryStreamRequest('plan_retry_factory_b'),
           authUserId,
           internalUserId,
           planId: 'plan_retry_factory_b',
